etc: rename project name: api -> springboot
